p2p service compiled
